+++ NEWS.md
@@ -1,1 +1,2 @@
 * Annotations relative to _data_ coordinates respect scale transformations.
+* Plots can be automatically unclipped to accommodate annotations extending beyond the plot boundaries. See `unclip()` and `set_ggtikz_unclip_hook()`.